friday/requirements.txt
ollama
sqlalchemy
python-dotenv
colorama
httpx
anthropic
PyQt6
faster-whisper
sounddevice
numpy
scipy
pykokoro
soundfile
pvporcupine
pyttsx3
